requirements.txt
mysql-connector-python
protobuf
python-dotenv
requests
sqlalchemy
twilio
promptlayer
psycopg2-binary
langchain
soundfile
openai
python-dotenv
numpy
matplotlib
plotly
pandas
scipy
sklearn
scikit-learn
